Add files via upload
An uncontrolled rectifier is an electronic circuit that converts alternating current (AC) into direct current (DC) without the ability to control or regulate the output voltage or current. Uncontrolled rectifiers typically use diodes as their main switching elements and are designed to provide a fixed and full-wave rectification of the incoming AC signal. They are simple and cost-effective devices, but they lack the ability to adjust the DC output voltage or current.